+++ unlock-js/src/__tests__/integration/walletServiceIntegration.test.js
@@ -35,6 +35,8 @@ jest.setTimeout(15000)
 // I wish we did not have to do this manually!
 const versionMappings = { v0: v0, v01: v01, v02: v02, v10: v10, v11: v11 }
 
+let accounts
+
 // Tests
 describe('Wallet Service Integration', () => {
   const versions = Object.keys(abis)
@@ -50,6 +52,7 @@ describe('Wallet Service Integration', () => {
       walletService = new WalletService({
         unlockAddress: initialization.to,
       })
+
       web3Service = new Web3Service({
         readOnlyProvider: provider,
         unlockAddress: initialization.to,
@@ -57,6 +60,7 @@ describe('Wallet Service Integration', () => {
       })
 
       await walletService.connect(provider)
+      accounts = await walletService.provider.listAccounts()
     })
 
     it('should yield true to isUnlockContractDeployed', done => {
@@ -77,11 +81,21 @@ describe('Wallet Service Integration', () => {
     describe.each(
       locks[versionName].map((lock, index) => [index, lock.name, lock])
     )('lock %i: %s', (lockIndex, lockName, lockParams) => {
-      let lock, lockAddress
+      let lock, lockAddress, lockCreationHash
 
       beforeAll(async () => {
-        lockAddress = await walletService.createLock(lockParams)
+        lockAddress = await walletService.createLock(
+          lockParams,
+          (error, hash) => {
+            lockCreationHash = hash
+          }
+        )
         lock = await web3Service.getLock(lockAddress)
+      })
+
+      it('should have yielded a transaction hash', () => {
+        expect.assertions(1)
+        expect(lockCreationHash).toMatch(/^0x[0-9a-fA-F]{64}$/)
       })
 
       it('should have deployed the right lock version', async () => {
@@ -119,15 +133,30 @@ describe('Wallet Service Integration', () => {
         )
       })
 
+      it('should have deployed a lock to the right owner', () => {
+        expect.assertions(1)
+        expect(lock.owner).toEqual(accounts[0]) // This is the default in walletService
+      })
+
       describe('updateKeyPrice', () => {
-        let oldKeyPrice, newPrice
+        let oldKeyPrice, newPrice, transactionHash
         beforeAll(async () => {
           oldKeyPrice = lock.keyPrice
-          newPrice = await walletService.updateKeyPrice({
-            lockAddress: lockAddress,
-            keyPrice: (parseFloat(oldKeyPrice) * 2).toString(),
-          })
+          newPrice = await walletService.updateKeyPrice(
+            {
+              lockAddress: lockAddress,
+              keyPrice: (parseFloat(oldKeyPrice) * 2).toString(),
+            },
+            (error, hash) => {
+              transactionHash = hash
+            }
+          )
           lock = await web3Service.getLock(lockAddress)
+        })
+
+        it('should have yielded a transaction hash', () => {
+          expect.assertions(1)
+          expect(transactionHash).toMatch(/^0x[0-9a-fA-F]{64}$/)
         })
 
         it('should have changed the keyPrice', () => {
@@ -143,9 +172,9 @@ describe('Wallet Service Integration', () => {
         let keyOwner, keyPurchaser
         let lockBalanceBefore
         let userBalanceBefore
+        let transactionHash
 
         beforeAll(async () => {
-          const accounts = await walletService.provider.listAccounts()
           keyPurchaser = accounts[0] // This is the default in walletService
           keyOwner = accounts[5]
           if (lock.currencyContractAddress === null) {
@@ -176,12 +205,22 @@ describe('Wallet Service Integration', () => {
             )
           }
 
-          tokenId = await walletService.purchaseKey({
-            lockAddress,
-            owner: keyOwner,
-            keyPrice: lock.keyPrice,
-          })
+          tokenId = await walletService.purchaseKey(
+            {
+              lockAddress,
+              owner: keyOwner,
+              keyPrice: lock.keyPrice,
+            },
+            (error, hash) => {
+              transactionHash = hash
+            }
+          )
           key = await web3Service.getKeyByLockForOwner(lockAddress, keyOwner)
+        })
+
+        it('should have yielded a transaction hash', () => {
+          expect.assertions(1)
+          expect(transactionHash).toMatch(/^0x[0-9a-fA-F]{64}$/)
         })
 
         it('should yield the tokenId', () => {
@@ -252,7 +291,96 @@ describe('Wallet Service Integration', () => {
         })
       })
 
-      describe('withdrawFromLock', () => {})
+      describe('withdrawFromLock', () => {
+        let lockBalanceBefore
+        let userBalanceBefore
+        let withdrawnAmount
+        let transactionHash
+        // TODO: support partial withdraws
+        // TODO: get the beneficiary address from the lock
+
+        beforeAll(async () => {
+          if (lock.currencyContractAddress === null) {
+            // Get the ether balance of the lock before the withdrawal
+            lockBalanceBefore = await web3Service.getAddressBalance(lockAddress)
+            // Get the ether balance of the beneficiary before the withdrawal
+            userBalanceBefore = await web3Service.getAddressBalance(lock.owner)
+          } else {
+            // Get the erc20 balance of the lock before the purchase
+            lockBalanceBefore = await web3Service.getTokenBalance(
+              lock.currencyContractAddress,
+              lockAddress
+            )
+            // Get the erc20 balance of the user before the purchase
+            userBalanceBefore = await web3Service.getTokenBalance(
+              lock.currencyContractAddress,
+              lock.owner
+            )
+          }
+
+          withdrawnAmount = await walletService.withdrawFromLock(
+            {
+              lockAddress,
+            },
+            (error, hash) => {
+              transactionHash = hash
+            }
+          )
+        })
+
+        it('should have yielded a transaction hash', () => {
+          expect.assertions(1)
+          expect(transactionHash).toMatch(/^0x[0-9a-fA-F]{64}$/)
+        })
+
+        it('should have withdrawn an non null amount', () => {
+          expect.assertions(1)
+          expect(withdrawnAmount).toEqual(lockBalanceBefore)
+        })
+
+        it('should decrease the balance by the withdrawn amount', async () => {
+          expect.assertions(1)
+          let lockBalanceAfter
+          if (lock.currencyContractAddress === null) {
+            // Get the ether balance of the lock before the withdrawal
+            lockBalanceAfter = await web3Service.getAddressBalance(lockAddress)
+          } else {
+            // Get the erc20 balance of the lock before the purchase
+            lockBalanceAfter = await web3Service.getTokenBalance(
+              lock.currencyContractAddress,
+              lockAddress
+            )
+          }
+          expect(parseFloat(lockBalanceAfter)).toEqual(
+            parseFloat(lockBalanceBefore) - parseFloat(withdrawnAmount)
+          )
+        })
+
+        it('should increase the balance of the beneficiary', async () => {
+          expect.assertions(1)
+          let beneficiaryBalanceAfter
+          if (lock.currencyContractAddress === null) {
+            // Get the ether balance of the beneficiary before the withdrawal
+            beneficiaryBalanceAfter = await web3Service.getAddressBalance(
+              lock.owner
+            )
+            // We should take gas paid into account... so the amount is larger than before
+            // but smaller than the sum of the previous balance and the amount in the lock
+            expect(parseFloat(beneficiaryBalanceAfter)).toBeGreaterThan(
+              parseFloat(userBalanceBefore)
+            )
+          } else {
+            // Get the erc20 balance of the user before the purchase
+            beneficiaryBalanceAfter = await web3Service.getTokenBalance(
+              lock.currencyContractAddress,
+              lock.owner
+            )
+            expect(parseFloat(beneficiaryBalanceAfter)).toEqual(
+              parseFloat(userBalanceBefore) + parseFloat(withdrawnAmount)
+            )
+          }
+        })
+      })
     })
   })
 })