+++ unlock-js/src/v01/createLock.js
@@ -6,8 +6,9 @@ import { UNLIMITED_KEYS_COUNT } from '../../lib/constants'
 /**
  * Creates a lock on behalf of the user, using version v01
  * @param {PropTypes.lock} lock
+ * @param {function} callback invoked with the transaction hash
  */
-export default async function(lock) {
+export default async function(lock, callback) {
   const unlockContract = await this.getUnlockContract()
   let maxNumberOfKeys = lock.maxNumberOfKeys
   if (maxNumberOfKeys === UNLIMITED_KEYS_COUNT) {
@@ -28,6 +29,10 @@ export default async function(lock) {
     transactionPromise,
     TransactionTypes.LOCK_CREATION
   )
+
+  if (callback) {
+    callback(null, hash)
+  }
 
   // Let's update the lock to reflect that it is linked to this
   // This is an exception because, until we are able to determine the lock address